% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addColumns.R
\name{addColumns}
\alias{addColumns}
\title{Add columns to data.frame}
\usage{
addColumns(trg, src, key, cols, after = NULL, before = NULL)
}
\arguments{
\item{trg, src}{target and source \code{data.frame}s}

\item{key}{key column(s). Either a single name, in which case it
applies to both source and target, or a string of the form
\code{"key1=key2"} where the first and second apply to target and
source, respectively. Multiple keys may be passed.}

\item{cols}{names of columns to copy to target. The columns can be
renamed by using the form \code{"new=old"}.}

\item{after, before}{string. Indicates after or before which of the
existing columns the new columns should be inserted. Default is
to append them to the right of existing columns.}
}
\value{
target data frame amended with columns.
}
\description{
Adds columns from another data.frame based on a key.
}
\examples{
data(CO2, package="datasets")

d <- aggr(CO2,"Plant", "mean.uptake=mean(uptake)")
addColumns(d, CO2, "Plant", c("Type","Treatment"))

## differnet key columns, and column renaming
d <- aggr(CO2,"p=Plant", "mean.uptake=mean(uptake)")
addColumns(d, CO2, p="Plant", c("name=Type","chilled=Treatment"))

}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
