% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine2D.R
\name{apply2D}
\alias{apply2D}
\title{Apply affine 2D transformation to (x,y) data}
\usage{
apply2D(m, x, y = NULL)
}
\arguments{
\item{m}{transformation matrix}

\item{x}{vector of x coordinates, or list with elements x and y when parameter y is NULL}

\item{y}{vector of y coordinates, or NULL when x is a list with elements x and y}
}
\value{
list with elements x and y
}
\description{
Transform (x,y) data passed either as separate vectors x and y or as list with
elements x and y.
}
\examples{
## none so far...
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
