% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axes.R
\name{bottom.axis}
\alias{bottom.axis}
\alias{left.axis}
\alias{top.axis}
\alias{right.axis}
\title{Plot axis labels}
\usage{
bottom.axis(...)

left.axis(...)

top.axis(...)

right.axis(...)
}
\arguments{
\item{...}{parameter passed to \code{axis}. See \code{?axis} for
details.}
}
\description{
\code{left.axis}, \code{bottom.axis}, \code{right.axis} and
\code{top.axis} simply plot the corresponding axes.
}
\details{
I always forget the numbering of the different axes, and therefore
introduced these convenience functions which simply pass through
all parameters to the base \code{axis} commend (but add the correct
axis number). A further difference is that by default (if
\code{las} is not specified), all labels are printed horizontally,
which I prefer for publication-quality figures.
}
\examples{
\dontrun{
## silly example...
plot(rnorm(10), rnorm(10), xlim=c(-3,3), ylim=c(-3,3),
     xaxt="n", yaxt="n", xaxs="i", yaxs="i")
xvec <- seq(-2, 2, by=2)
left.axis(at=xvec)
bottom.axis(at=xvec)
}
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
