% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catenary.R
\name{catenary}
\alias{catenary}
\alias{catenary_normal}
\title{Catenary curves}
\usage{
catenary(x, a = NULL, dx = 1, y1 = 0, y2 = 0, smax = NULL)

catenary_normal(x, a = NULL, dx = 1, y1 = 0, y2 = 0, smax = NULL)
}
\arguments{
\item{x}{location at which y is to be determined. Should be between
\code{0} and \code{dx}.}

\item{a}{shape parameter}

\item{y1, y2, dx}{y values and horizontal distance between end points
of the caternary. The end points thus are (0,y1) and (dx,y2).}

\item{smax}{alternative shape parameter indicating the maximum sag}
}
\value{
\code{catenary} returns a list with coordinates along the
    caternary (\code{x}, \code{y}), the lowest point (\code{xmin},
    \code{ymin}), the point of maximum sag (\code{xc}, \code{yc}),
    the shape parameter \code{a}, and the maximum sag \code{smax}.
    \code{caternary_normal} returns points along the secant
    (\code{xline},\code{yline}), and a second set of points on the
    caternary (\code{x},\code{y}). The connections between both
    point sets are normal to the secant. A last element \code{dist}
    returns the length of these connections.
}
\description{
\code{catenary} constructs a catenary given some
parameter. \code{catenary_normal} returns vectors normal to the
secant between the catenary's end points to the catenary.
}
\details{
A catenary has the general form \eqn{y = a cosh(x/a)}, with shape
parameter \code{a}. These functions re-parameterize this general
form in terms of y values \code{y1} and \code{y2} at the left and
right end of the caternary, and the horizontal distance between end
points \code{dx}. As alternative to the shape parameter \code{a},
the maximum sag \code{smax} along the catenary can be passed (in
other words: the maximum vertical distance between the direct line
connecting the end points and the catenary).

For simplicity of implementation, the re-parameterization of the
catenary in terms of maximum sag \code{smax} instead of \code{a} is
done with an interative search using the Newton
algorithm. Similarly, the normals to the secant are determined iteratively.
}
\examples{
\dontrun{
dx <- 2
y1 <- 2
y2 <- 3
r <- catenary(x=seq(0, dx, length=30), dx=dx, y1=y1, y2=y2, smax=.7)
plot(r, type="l", asp=1, las=1, ylab="y", xlab="x")
lines(c(0, dx), c(y1, y2), lty=2, col="gray")
abline(h=r$ymin, col="red")
text(1.5, r$ymin, pos=3, "ymin", col="red")
abline(v=c(0, dx), col="#80808080")
points(r$xc, r$yc, col="blue")
arrows(r$xc, r$yc, r$xc, r$yc+r$smax, col="orange", length=.1, code=3)
text(r$xc, r$yc+0.5*r$smax, srt=90, pos=2, col="orange", "smax")

# add the normals to the secant
r <- catenary_normal(x=seq(0, dx, length=10), dx=dx, y1=y1, y2=y2, smax=.7)
arrows(r$xline, r$yline, r$x,r$y, length=.1, col="blue")
}

}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
