% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cloglog.R
\name{cloglog}
\alias{cloglog}
\alias{invcloglog}
\title{Complementary log-log link function and its inverse}
\usage{
cloglog(p)

invcloglog(eta)
}
\arguments{
\item{x}{A vector, list or matrix containing character or numeric data, possibly as factor.}
}
\value{
A numeric vector, list or matrix containing the converted data.
}
\description{
\code{cloglog} will return the complementary log log function, whereas
\code{invcloglog} will invert it.
\code{cloglog} will yield -Inf and Inf for arguments 0 and 1, or NaN for values
outside this range. Conversely, \code{invcloglog} handles inputs of -Inf and Inf.
}
\examples{
library(pascal)
x <- factor(3:5)
x
## [1] 3 4 5
## Levels: 3 4 5
as.numeric(x)
## [1] 1 2 3
safen(x)
## [1] 3 4 5
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}

