% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combn2.R
\name{combn2}
\alias{combn2}
\title{Combinations of n elements, taken m at a time, with optional replacement}
\usage{
combn2(x, m, replace = FALSE, FUN = NULL)
}
\arguments{
\item{x}{vector with data from which combinations are created}

\item{m}{number of elements to choose#'}

\item{replace}{logical indicating whether combinations are created
by drawing with replacement, i.e. whether duplicate elements are
allowed}

\item{FUN}{optional function to be applied to combinations}
}
\value{
Matrix with combinations in columns, or result of FUN
applied to columns if FUN is provided
}
\description{
This function is similar to \code{combn} but allows for
combinations resulting from drawing with replacement.
}
\details{
To keep implementation simple, \code{combn2} internally uses
recursion and is slower than \code{combn}, but this should not be a
problem for `reasonably' sized data. Recursion depth is m-2.
}
\examples{
combn2(LETTERS[1:3],2) 
##     [,1] [,2] [,3]
## [1,] "A"  "A"  "B" 
## [2,] "B"  "C"  "C"

combn2(LETTERS[1:3],2,replace = TRUE)
##     [,1] [,2] [,3] [,4] [,5] [,6]
## [1,] "A"  "A"  "A"  "B"  "B"  "C" 
## [2,] "A"  "B"  "C"  "B"  "C"  "C"

combn2(LETTERS[1:3],2,replace = TRUE,FUN=function(x) paste(x,collapse=":"))
## [1] "A:A" "A:B" "A:C" "B:B" "B:C" "C:C"

}
\seealso{
\code{\link{combn}}
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
