% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cot.R
\name{cot}
\alias{cot}
\alias{cotpi}
\title{Cotangent}
\usage{
cot(z)

cotpi(z)
}
\arguments{
\item{z}{argument in radians, or multiple of pi}
}
\value{
cotangent of z
}
\description{
\code{cot} calculates the cotangent. It is better than computing
1/tan(x) like done in \code{pracma} because it does not fail for
\code{pi/2}.
}
\examples{
cot(0.25*pi)

}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
