% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/current_time.R
\name{current_time}
\alias{current_time}
\title{Return the current date and time, nicely formatted}
\usage{
current_time(fmt = "\%d-\%b-\%Y \%H:\%M:\%S")
}
\arguments{
\item{fmt}{A string specifying the format of date and time. Defaults to \code{\%d-\%b-\%Y \%H:\%M:\%S}.
Check \code{\link{strftime}} for format specifications.}
}
\value{
Current data and time as character string, formatted as specified
}
\description{
Convenience function that returns date and time formatted as I like them
}
\examples{
current_time()
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
\seealso{
\code{\link{Sys.time}}, \code{\link{strftime}}
}

