% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resDivider.R
\docType{data}
\name{E12}
\alias{E12}
\title{Resistor E12 series values}
\format{
An vector of class \code{"numeric"}
}
\usage{
data(E12)
}
\description{
Resistor E12 series values, with R in [1,10[.
}
\examples{
data(E12)
E12
}
\keyword{datasets}
