% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resDivider.R
\docType{data}
\name{E24}
\alias{E24}
\title{Resistor E24 series values}
\format{
An vector of class \code{"numeric"}
}
\usage{
data(E24)
}
\description{
Resistor E24 series values, with R in [1,10[.
}
\examples{
data(E24)
E24
}
\keyword{datasets}
