% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resDivider.R
\docType{data}
\name{E48}
\alias{E48}
\title{Resistor E48 series values}
\format{An vector of class \code{"numeric"}}
\usage{
data(E48)
}
\description{
Resistor E48 series values, with R in [1,10[.
}
\examples{
data(E48)
E48
}
\keyword{datasets}
