% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{ellipse}
\alias{ellipse}
\title{Confidence ellipse}
\usage{
ellipse(data = NULL, cov = NULL, center = NULL, level = 0.95,
  npoints = 100)
}
\arguments{
\item{data}{data frame or matrix with x and y values in first and
second column}

\item{cov}{covariance matrix of bivariate data}

\item{center}{center of ellipsoid}

\item{level}{confidence interval to draw (default = 0.95)}

\item{npoints}{number of points returned for the confidence ellipse
(default 100)}
}
\value{
list with parameters specifying the confidence
    ellipse. In particular, \code{ellipse} is a list of coordinates
    delineating the ellipse, and \code{ma1} and \code{ma2} the
    end points of the major axes. \code{scale} are the standard
    deviations of x and y.  \code{center} is the center of the ellipsoid.
    \code{r} is the correlation of x and y.
}
\description{
Calculate a confidence ellipse for bivariate data
}
\details{
This function is useful to calculate confidence ellipses and their
major axes into ordination plots. Data can be provided either as
set of (x,y) values in a data frame or matrix (argument
\code{data}), or as covariance matrix \code{cov} and mean x and y
values \code{center}. \code{level} specifies the confidence
interval to draw.
}
\note{
Part of the code is taken and adapted from the \code{ellipse} package,
    for my own purpouses.
}
\examples{
p <- rnorm(100,sd=1)
d <- data.frame(x=30+p+rnorm(100,sd=1), y=20-.5*p+rnorm(100,sd=2))
e <- ellipse(d)
plot(d$x,d$y,asp=1)
lines(e$ellipse, col="red")
lines(e$ma1,col="red")
lines(e$ma2,col="red")
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
\keyword{misc,}
\keyword{utilities}
