% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitScreen.R
\name{emptyPlot}
\alias{emptyPlot}
\title{Provides an empty plot}
\usage{
emptyPlot(xlim = c(0, 1), ylim = c(0, 1), asp = NA, xpd = NA,
  xlab = "", ylab = "", bty = "n", xaxs = "i", yaxs = "i",
  xaxt = "n", yaxt = "n", ...)
}
\arguments{
\item{xlim, ylim}{extents of plot. Defaults to c(0,1) for both axis.}

\item{asp}{aspect ratio. Defaults to NA.}

\item{xpd}{Defauts to NA. Check `?par` for details.}

\item{xlab, ylab}{axis labels (default: empty)}

\item{bty}{box type. Defaults to none ('n'; see `?par` for details)}

\item{xaxs, yaxs, xaxt, yaxt}{axes characteristics. Defaults to 'i'
and 'n'. See `?par` for details.}

\item{...}{Additional parameters passed to `plot`}
}
\description{
This function
}
\details{
This function provides a parametric interface to `split.screen` for
regularly spaced and sized panels. The parameters are self-explanatory.
}
\examples{
emptyPlot()

}
\seealso{
\code{\link{par}}, \code{\link{plot}}
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
