% Generated by roxygen2 (4.0.2): do not edit by hand
\name{escape}
\alias{escape}
\title{Escape characters in strings}
\usage{
escape(x, to.escape = ":")
}
\arguments{
\item{x}{Vector, list or matrix of character strings}

\item{to.escape}{String containing the characters to escape (default ':')}
}
\value{
Vector, list or matrix of string with the corresponding characters escaped
}
\description{
Replace selected characters in vectors, lists, or matrices by \\xxxx, where xxxx is their hexadecimal representation.
Note that the backslash is always escaped, since it is used as escape character.
}
\examples{
r<-escape(list("A:B","C:D"))
r
## [[1]]
## [1] "A\\\\003aB"
##
## [[2]]
## [1] "C\\\\003aD"
unescape(r)
## [[1]]
## [1] "A:B"
##
## [[2]]
## [1] "C:D"
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
\seealso{
\code{\link{unescape}}
}

