% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite.only.R
\name{finite.only}
\alias{finite.only}
\title{Return finite elements of a list or vector}
\usage{
finite.only(x)
}
\arguments{
\item{x}{vector of data}
}
\value{
vector with non-finite (NaN, Inf) and missing data (NA) removed.
}
\description{
Convenience function that removes non-finite elements (i.e. NA, NaN, Inf, -Inf) from a list or vector
}
\details{
Character vectors are first converted to numeric values. Similarly, factors are first converted to 
their character representation and then to numeric values.
}
\examples{
x <- 1:5
x[3]<-NA
x[4]<-1/0
x
## [1]   1   2  NA Inf   5
finite.only(x)
## [1] 1 2 5
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}

