% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDist.R
\name{getDist}
\alias{getDist}
\title{Extract distance from "dist" object}
\usage{
getDist(d, i, j)
}
\arguments{
\item{d}{\code{dist} object}

\item{i}{item i. Must be either a label or an integer index; can be a vector.}

\item{j}{item j, as described for i.}
}
\value{
distance(s) between \code{i} and \code{j}
}
\description{
Objects of class \code{dist} are triangular
matrices. \code{getDist} extracts distances between specific pairs
of items.
}
\details{
There is, as of today (July 2015), no simple way to extract
individual distances from \code{dist} objects. The present
function indices the \code{dist} object, given two labels
specifying the item to extract.
}
\examples{
data(USJudgeRatings)
d <- dist(USJudgeRatings)
getDist(d,c("PASKEY,L.","SHEA,D.M.","LEVINE,I."),rep("LEVINE,I.",3))
## [1] 1.56 2.35 0.00
}
\seealso{
\code{\link{dist}}
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
