% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getObjects.R
\name{getObjects}
\alias{getObjects}
\title{Get objects of a certain class}
\usage{
getObjects(cls = "function")
}
\arguments{
\item{cls}{class type (default \code{function})}
}
\value{
list of objects of the specified class
}
\description{
This function returns the objects defined in the current frame that
are of a given class (by default functions)
}
\examples{
library(pascal)
a<-1
b <- function(x) { x }
getObjects("numeric");
getObjects("function");
## remove all objects that are not functions
ls()
rm(list=setdiff(ls(),getObjects()));
ls()
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}

