% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getResolution}
\alias{getResolution}
\title{Get current screen resolution}
\usage{
getResolution()
}
\value{
A list with elements \code{width} and \code{height}
}
\description{
This function calls \code{xrandr} to get the resolution of the display in use.
If the system is not linux (and xrandr not available), it returns a width of 1280 and a height of 1024 pixels.
This could be fixed by adding platform-specific code.
}
\examples{
getResolution()
## $width
## [1] 1366
##
## $height
## [1] 768
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}

