% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfx2Dclass.R
\docType{class}
\name{gfx2D-class}
\alias{gfx2D-class}
\alias{gfx2D}
\alias{show,gfx2D-method}
\alias{transformation<-,gfx2D-method}
\alias{transformation<-}
\alias{transformation,gfx2D-method}
\alias{transformation}
\title{gfx2D class and initialization

The class \code{gfx2D} is used for plotting, applying to all
drawing operations a geometric (2D-affine) transformation that is
stored as $3 x 3$ matrix.

The function \code{gfx2D} creates an object of this class and
initializes its transformation matrix. The transformation
operations can be passed as arguments (scaling: \code{sx},
\code{sy}; rotation: \code{rot}; translation: \code{tx},
\code{ty}).

transformation (get, set)

drawing operations described separately in ?plot2D}
\usage{
gfx2D(tx = c(0, 0), ty = tx[2], sx = c(1, 1), sy = sx[2], rot = 0)

\S4method{show}{gfx2D}(object)

\S4method{transformation}{gfx2D}(object) <- value

\S4method{transformation}{gfx2D}(object)
}
\arguments{
\item{tx, ty}{translation, given either as 2-element vector
\code{tx}, or separately as \code{tx} and \code{ty}.}

\item{sx, sy}{scaling in x and y direction,  given either as 2-element vector
\code{sx}, or separately as \code{sx} and \code{sy}.}

\item{rot}{rotation, in radians.}

\item{object}{gfx2D object}

\item{value}{transformation matrix

\code{transformation(object)} displays the transformation matrix
associated with the \code{gfx2D}-object. The transformation matrix
can be set with \code{transformation(object) <- m}, where code{m}
is a 3 x 3 matrix.  The scaling, rotation, and translation embodied
in \code{m} are extracted and stored separately in the object.}
}
\value{
gfx2D-object
}
\description{
gfx2D class and initialization

The class \code{gfx2D} is used for plotting, applying to all
drawing operations a geometric (2D-affine) transformation that is
stored as $3 x 3$ matrix.

The function \code{gfx2D} creates an object of this class and
initializes its transformation matrix. The transformation
operations can be passed as arguments (scaling: \code{sx},
\code{sy}; rotation: \code{rot}; translation: \code{tx},
\code{ty}).

transformation (get, set)

drawing operations described separately in ?plot2D

\code{show(object)} displays the transformation matrix and possibly
other options associated with the \code{gfx2D} object.
}
\examples{

g <- gfx2D()

transformation(g) <- rotate2D(pi/4)

}
\seealso{
plot2D,affine2D
}
