% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse.R
\name{glimpse}
\alias{glimpse}
\title{Print an excerpt of some data}
\usage{
glimpse(..., len = 5, pre = 1, post = 0, between = 1)
}
\arguments{
\item{...}{data to be printed}

\item{len}{number of elements to be printed of head and tail
section of the data.}

\item{pre, post, between}{number of empty lines to print before and after
the output, and in between elements if there are multiple.}
}
\description{
This function allows to print an excerpt of some data structures
(currently data.frame, vector, list).  The first and last 'len'
elements are printed, with '...' in between the head and tail
parts.
}
\details{
I use this function in status reports or log files to show what
type of output is produced.
}
\examples{
data(CO2)
a <- 1:20
b <- as.list(LETTERS[1:20])
glimpse(CO2,a,b,len=3)
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
