% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corner.label.R
\name{gline}
\alias{gline}
\title{Add a line to a plot using grid coordinates}
\usage{
gline(x, y, frame = "plot", ...)
}
\arguments{
\item{x, y}{Coordinates of the labels, or a string containing a
combination of dimensions and units
(e.g. "2native+2.5lines"). Coordinate vectors are recycled to
equal length if necessary, allowing to pass just a single x or y value
for vertical and horizontal lines, respectively.}

\item{frame}{Either of 'inner', 'figure', or 'plot'. Indicates the
area that is to be labelled. 'inner' is the entire page area
except the margins. 'figure' is the figure area. 'plot' is the
area used for plotting (inside the axis). Defaults to 'plot'.}

\item{...}{Extra parameters passed when calling
\code{grid.line}. Check \code{?gpar} for options.}
}
\description{
\code{gline} draws a line with coordinates that can be specified in
different units.
}
\details{
This function facilitates the decoration of plots. Using
\code{grid}-graphics, lines are drawn using coordinates that can be
specified in different units.
}
\examples{
\dontrun{
plot(NULL, xlim=c(0,2), ylim=c(1,3))
gline(c("0.5native","1.5native"), "-1lines", col="blue",lwd=2)
}
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
