% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmgcd.R
\name{greatestCommonDenominator}
\alias{greatestCommonDenominator}
\title{Greatest common denominator}
\usage{
greatestCommonDenominator(a, b)
}
\arguments{
\item{a}{integer}

\item{b}{integer}
}
\value{
greatest common denominator
}
\description{
Given two integers, determines the greates common denominator.
}
\examples{
library(pascal)
greatestCommonDenominator(15,9)
}
\references{
This functions are adapted from library \code{pracma}
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
