% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplots.R
\name{groupSpace}
\alias{groupSpace}
\title{Add spacing between bars in bar plot}
\usage{
groupSpace(d, cols, equal = FALSE)
}
\arguments{
\item{d}{data frame with data}

\item{cols}{the names of the columns that define the groups.}

\item{equal}{logical indicating whether the number of transitions
(when multiple columns are specified) should be returned
instead of a simple vector of \code{1} or \code{0}.}
}
\value{
A vector with values indicating transitions.
}
\description{
\code{groupSpace} is a helper function to add space between the
bars of a bar plot.
}
\details{
Given a \code{data.frame} and the names of columns in it, it
determines transitions between the groups these columns define and
returns 1 for a transition, or 0 otherwise. These values can be
passed to the \code{space} argument of \code{barplot}.
}
\examples{
data(CO2)

d <- aggregate(uptake ~ Plant + Type + Treatment, data=CO2, FUN = mean)

barplot(d$uptake, names.arg = d$Plant,
        space = groupSpace(d, c("Type","Treatment")))

barplot(d$uptake, names.arg = d$Plant,
        space = groupSpace(d, c("Type","Treatment"), equal = TRUE))
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
