% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heading.R
\name{headingtrailer}
\alias{heading}
\alias{headingtrailer}
\alias{simple.heading}
\alias{simple.trailer}
\alias{trailer}
\title{Functions to print nice headings and trailer/footers}
\usage{
heading(txt = NULL, width = 80, right = FALSE, bottom = FALSE,
  char = "#", spc = "  ", side.char = char, corner.char = rep(char, 4),
  post = 1, pre = 0, center = FALSE)

trailer(txt = NULL, width = 80, right = FALSE, char = "#", spc = "  ",
  side.char = char, corner.char = rep(char, 4), post = 1, pre = 0,
  center = FALSE)

simple.heading(txt = NULL, width = 80, char = "#", side.char = char,
  corner.char = rep(char, 4), spc = "  ", pre = 0, post = 1,
  center = TRUE)

simple.trailer(txt = NULL, width = 80, char = "#", side.char = char,
  corner.char = rep(char, 4), spc = "  ", pre = 0, post = 1,
  center = TRUE)
}
\arguments{
\item{txt}{A vector or list of strings that are to be printed in the heading or trailer.
Each element is printed on a separate line}

\item{width}{Width of the heading or trailer (default: 80 characters)}

\item{right}{TRUE indicates that the right border of the frame is to be printed (default: FALSE)}

\item{bottom}{TRUE indicates that the bottom border of the frame is to be printed (default: FALSE)}

\item{char}{character to draw to horizontal parts of the frame (default is '#')}

\item{spc}{string used as spacing before and after (left and right) of text lines.
Usually, these are a series of spaces, but it is possible to use different characters.
Default are two spaces.}

\item{side.char}{character to draw to vertical parts of the frame (default is the same as \code{char})}

\item{corner.char}{character vector to draw the corners of frames. The elements of the vector are
(in this order) top left, top right, bottom left, and bottom right.)
Default is to use the same character as for horizontal lines.}

\item{post}{Number of empty lines to print after the header/trailer. Default is 1.}

\item{pre}{Number of empty lines to print before the header/trailer. Default is 0.}

\item{center}{TRUE indicates centered text, FALSE indicates left-aligned text.}
}
\value{
These functions don't return a value
}
\description{
Functions to print nice headings and trailer/footers
}
\examples{
heading("Results")
## ##############################
## #
## #  Results
## #
heading("Results",center=TRUE,bottom=TRUE,right=TRUE,char="-",side.char="|",corner.char=c("/","\\\\"),width=30)
## /----------------------------\\
## |                            |
## |          Results           |
## |                            |
## \\----------------------------/
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}

