% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexof.R
\name{indexof}
\alias{indexof}
\title{Find column index in data frame}
\usage{
indexof(d = NULL, col = NULL)
}
\arguments{
\item{d}{data frame}

\item{col}{column name}
}
\value{
indices of columns with the specified names
}
\description{
Deprecated -- is just there for compatibility with old code
and will eventually be removed
}
\seealso{
\code{\link{which}}
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
