% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isect.R
\name{isect}
\alias{isect}
\title{Intersect two lines}
\usage{
isect(x1, y1, x2, y2, segment = T)
}
\arguments{
\item{x1}{Vector or two specifying the end point x coordinates of first line}

\item{y1}{Vector or two specifying the end point y coordinates of first line}

\item{x2}{Vector or two specifying the end point x coordinates of second line}

\item{y2}{Vector or two specifying the end point y coordinates of second line}

\item{segment}{Logical determining whether the the lines are segments ending
               at the end points specified, or whether the lines have infinite length.}
}
\value{
Vector with coordinates of intersection \code{x,y}. These coordinates equal
        NA if the segment option is specified and the segments don't intersect
}
\description{
Calculate the point (x,y) at which two lines intersect. Note that the specification
or the line is organised differently than in e.g. \code{arrows}.
}
\examples{
isect(c(0,1),c(0,1),c(0,1),c(1,0))
## [1] 0.5 0.5
isect(c(0,1),c(0,1),c(0,3),c(1,2))
## [1] NA NA
isect(c(0,1),c(0,1),c(0,3),c(1,2),segment=FALSE)
## [1] 1.5 1.5
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
\keyword{misc,}
\keyword{utilities}

