% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/killfactors.R
\name{killfactors}
\alias{killfactors}
\title{Remove factors property from data frame columns}
\usage{
killfactors(d, restore.numerics = FALSE, cols = seq_along(names(d)))
}
\arguments{
\item{d}{Source data frame containing the data set to aggregate}

\item{restore.numerics}{Logical determining whether numeric columns
    are converted to numeric instead of the default character type
    (default FALSE). A column is considered numeric when all values
    except NA can be converted to valid numeric data including NaN
    and Inf.}

\item{cols}{Vector determining which columns should be inspected
    (default: all). Columns can be specified by index or name.}
}
\description{
\code{killfactors} converts factors in data frames to regular
numeric of character vectors.  R automatically converts strings to
factors when data frames are created. This is not always desired,
and \code{killFactors} will ``undo'' this behaviour.
}
\examples{
d <- data.frame(c1=LETTERS[1:5],c2=factor(1:5),c3=1:5)
str(d)
## 'data.frame':	5 obs. of  3 variables:
##  $ c1: Factor w/ 5 levels "A","B","C","D",..: 1 2 3 4 5
##  $ c2: Factor w/ 5 levels "1","2","3","4",..: 1 2 3 4 5
##  $ c3: int  1 2 3 4 5
str(killfactors(d))
## 'data.frame':	5 obs. of  3 variables:
##  $ c1: chr  "A" "B" "C" "D" ...
##  $ c2: chr  "1" "2" "3" "4" ...
##  $ c3: int  1 2 3 4 5
str(killfactors(d,restore.numerics = TRUE))
## 'data.frame':	5 obs. of  3 variables:
##  $ c1: chr  "A" "B" "C" "D" ...
##  $ c2: num  1 2 3 4 5
##  $ c3: int  1 2 3 4 5
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}

