% Generated by roxygen2 (4.0.2): do not edit by hand
\name{logstc}
\alias{logstc}
\title{Logistic function}
\usage{
logstc(x, a = 1, b = 1, tau = 1)
}
\arguments{
\item{x}{points where the function is to be evaluated}

\item{a}{scaling factor, i.e. maximum value the function reaches}

\item{b}{x coordinate of point of inflection}

\item{tau}{coefficient indicating the growth in the exponential part of the curve,
i.e. the curve approximates \code{exp(tau*x)}}
}
\value{
function evaluated at locations passes in \code{x}
}
\description{
This is a three-parameter sigmoidal (logistic) function.
}
\details{
The function is defined as \code{f(x,a,b,tau)=a*(1-1/[1+exp(tau*(t-b))])}.
}
\examples{
x<-0:100
plot(-1,-1,xlim=range(x),ylim=c(0,1),xlab="x",ylab="logstc")
for(tau in c(.1,.5,1)) {
   lines(x,logstc(x,1,50,tau))
}
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
\keyword{misc,}
\keyword{utilities}

