% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoryuse.R
\name{memused}
\alias{memused}
\title{Display memory use statistcs}
\usage{
memused(
  print = TRUE,
  total = TRUE,
  objs = TRUE,
  top = 10,
  base2 = FALSE,
  env = globalenv()
)
}
\arguments{
\item{print}{logical, indicating whether results should be printed
(\code{TRUE}, default) or returned (\code{FALSE})}

\item{total}{logical, indicating whether the total memory used
should be reported. This is the data returned by a call to gc()}

\item{objs}{logical, indicating whether the memory use of the
\code{top} largest objects in environment \code{env} should be
listed.}

\item{top}{number of objects that should be included in the
list. The remaining are listed as totals at the end of the
table.}

\item{base2}{logical, indicating whether memory units are according
    to SI (base 1000) or base 2. Note that, for example,
    \code{gc()} reports 1MB=2^20=104856 bytes (default
    \code{FALSE}).
#'}

\item{env}{The environment from which objects should be
reported. Defaults to the global environment.}
}
\value{
numeric value with total memory used as reported by gc(),
    in bytes (\code{total=TRUE}), a \code{data.frame} with object
    memory use (\code{objs=TRUE}), or a list of both.
}
\description{
Displays (or returns) total memory use after garbage collection and
the largest R objects present in an environment (by default: the
global environment).  This function may be useful to identify
memory hogs.
}
\details{
Sometimes it is useful to get a quick overview of the largest
objects in a particular environment.  \code{memused} does exactly
this. By default, the largest objects in the global environment are
shown. If run within a function, one may wish to show objects in
the local environment instead.  See the examples below to learn
more about its use.
}
\examples{
a <- matrix(rnorm(1e6),1000,1000)
b <- LETTERS[1:26]
c <- integer(100000)
memused()
## -----------------  memory use  -------------------
## Total:  135.4 MByte
## Largest objects in R_GlobalEnv:
##              class      type   bytes percent       hsize
## a     matrix,array    double 8000216  95.22\%   8.0 MByte
## c          integer   integer  400048   4.76\% 400.0 KByte
## b        character character    1712   0.02\%  1712 Bytes
## TOTAL                  TOTAL 8401976 100.00\%   8.4 MByte

fun <- function(x) { a<-3; memused(env=environment()) }
fun(1:100)
## -----------------  memory use  -------------------
## Total:  135.4 MByte
## Largest objects in :
##         class    type bytes percent       hsize
## x     integer integer   448  88.89\%   448 Bytes
## a     numeric  double    56  11.11\%    56 Bytes
## TOTAL           TOTAL   504 100.00\%   504 Bytes

}
\seealso{
\code{\link{gc}}
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
