% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modalpha.R
\name{modalpha}
\alias{modalpha}
\title{Modify alpha channel of color}
\usage{
modalpha(x, k = NULL, alpha = NULL)
}
\arguments{
\item{x}{vector of colors (named colors, or in format #RRGGBB or
#RRGGBBAA)}

\item{k}{factor by which alpha channel value is changed (1.0 = no change).}

\item{alpha}{new alpha channel value, in range [0,255]}
}
\value{
vector with modified color values
}
\description{
Given a vector of colors, modify the alpha channel value by a given
factor or change it to a certain value.
}
\details{
This function deals with names colors (e.g. 'red') and hexadecimal
colors with and without alpha channel (e.g. '#ff000080' or
'#ff0000'). If a change factor 'k' is given, the alpha channel
value is modified by this factor. If a new 'alpha' value is given,
the alpha channel is set to this value.
}
\examples{

colrs <- c('red', '#ff000080', '#0000ff')

modalpha(colrs, alpha=0x80)
## [1] "#ff000080" "#ff000080" "#0000ff80"
modalpha(colrs, k=0.33)
## [1] "#ff000054" "#ff00002a" "#0000ffNA"

}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
