% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveColumns.R
\name{moveColumns}
\alias{moveColumns}
\title{Move columns in data.frame}
\usage{
moveColumns(d, cols, after = NULL, before = NULL)
}
\arguments{
\item{d}{\code{data.frame}}

\item{cols}{names of columns to move around in the data frame. The
columns can be renamed by using the form \code{"new=old"}.}

\item{after, before}{string. Indicates after or before which of the
existing columns the columns should be placed.}
}
\value{
data frame with different column order
}
\description{
Move columns in data frame to another position, possibly renaming
them.
}
\examples{

data(CO2, package="datasets")

CO2[1,]
##   Plant   Type  Treatment conc uptake
## 1   Qn1 Quebec nonchilled   95     16

d <- moveColumns(CO2, c("CO2=conc", "rate=uptake"), after="Plant")
d[1,]
##   Plant CO2 rate   Type  Treatment
## 1   Qn1  95   16 Quebec nonchilled

}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
