% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAsafes.R
\name{mean_NAsafe}
\alias{max_NAsafe}
\alias{mean_NAsafe}
\alias{median_NAsafe}
\alias{min_NAsafe}
\alias{sd_NAsafe}
\alias{se_NAsafe}
\alias{var_NAsafe}
\title{Convenience functions excluding NAs}
\usage{
mean_NAsafe(x)

se_NAsafe(x)

var_NAsafe(x)

sd_NAsafe(x)

median_NAsafe(x)

min_NAsafe(x)

max_NAsafe(x)
}
\arguments{
\item{x}{A data vector possibly containg NAs}
}
\value{
The result of the respective summary function, called with \code{na.rm=TRUE}
}
\description{
These functions call standard functions with the extra argument na.rm=TRUE.
These are shortcuts that generally could be replaced by passing anonymous functions
as arguments (see examples below).
}
\examples{
#
testDF<-data.frame(by=rep(LETTERS[1:2],each=4),y=c(1,3,2,9,4,4,NA,1))
aggregate(testDF$y,by=list(by=testDF$by),FUN=mean)
#    by    x
#  1  A 3.75
#  2  B   NA
aggregate(testDF$y,by=list(by=testDF$by),FUN=mean_NAsafe)
#    by    x
#  1  A 3.75
#  2  B 3.00
aggregate(testDF$y,by=list(by=testDF$by),function(x) { mean(x,na.rm=TRUE) } )
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
\seealso{
\code{\link{mean}} \code{\link{var}} \code{\link{se}} \code{\link{sd}} \code{\link{median}} \code{\link{min}} \code{\link{max}}
}

