% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAtozero.R
\name{NAtozero}
\alias{NAtozero}
\title{Replace NA and NaN with zero}
\usage{
NAtozero(x)
}
\arguments{
\item{x}{vector of data}
}
\value{
vector with non-finite (NaN, Inf) and missing data (NA) replaced by zero
}
\description{
Convenience function that replaces NA and NaN in vectors, lists, and matrices by zero
}
\details{
Note that +Inf/-Inf values are preserved.
}
\examples{
x <- 1:5
x[3]<-NA
x[4]<-1/0
x
## [1]   1   2  NA Inf   5
NAtozero(x)
## [1] 1 2 0 0 5
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}

