% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nice_path_diagram.R
\name{niceDiag}
\alias{niceDiag}
\title{Generate path diagram using graphViz}
\usage{
niceDiag(sem.fit, pdfFile = "fit.pdf", scaling.factor = 1,
  show.values = TRUE, boxtext = NULL, dispBIC = FALSE)
}
\arguments{
\item{sem.fit}{sem object}

\item{pdfFile}{file name of outout}

\item{scaling.factor}{xxx}

\item{show.values}{xxx}

\item{boxtext}{xxx}

\item{dispBIC}{xxx}
}
\description{
Deprecated -- is just there for compatibility with old code
and will eventually be removed
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}

