% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.hist.R
\name{panel.hist}
\alias{panel.hist}
\title{Histogram panel for \code{pairs}}
\usage{
panel.hist(x, ...)
}
\arguments{
\item{x}{values passed from \code{\link{pairs}}}

\item{...}{additional arguments passed to drawing command
(\code{\link{rect}})}
}
\description{
\code{pairs} offers panel functions, but the one to display a
histogram in the diagonal elements is not predefined (althouth
there is an example in the help).
}
\details{
This function implements such a function, and also produced
reasonable histograms for log-transformed scales.
}
\examples{
data(iris, package = "datasets")
iris[, 1:4]

##    Sepal.Length Sepal.Width Petal.Length Petal.Width
## 1           5.1         3.5          1.4         0.2
## 2           4.9         3.0          1.4         0.2
## 3           4.7         3.2          1.3         0.2
## 4           4.6         3.1          1.5         0.2

pairs(iris[, 1:4], upper.panel = NULL, diag.panel = panel.hist)

pairs(iris[, 1:4], log = 1:4, upper.panel = NULL, diag.panel = panel.hist)

## Probably not what you want, but possible:
pairs(iris[, 1:2], log = 1:2, diag.panel = function(x) panel.hist(x, col = "green", border = "magenta", lwd = 2))

}
\seealso{
\code{\link{pairs}}
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
