% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_color.R
\name{pick_color}
\alias{pick_color}
\title{Pick named colors}
\usage{
pick_color(hue = NULL, minsat = 0.2)
}
\arguments{
\item{hue}{hue of the color range to be shown. This may be a word
(e.g. "blue") or a numeric value in the range [0, 1[.}

\item{minsat}{minimum saturation of colors that are to be shown}
}
\value{
Data frame with color properties
}
\description{
\code{pick_color} shows named colors and allows to select these by
left-clicking on them with the mouse. Right-clicking terminates the
color selection.
}
\details{
It is possible to only show colors in a certain hue range. This is
done by passing a \code{hue} argument, which is either numeric
(range 0 to 1) or a text that is searched within the known color
names. The hue value is then calculated as the average hue of the
names that match the search.  Then, all colors within a hue range
of [-0.1, 0.1] around that mean value are presented.

Because there are 100 named grey values, only the ones in intervals
of 10 are shown (e.g. gray10, gray20 to gray90). Also, by default,
only colors with a minimum saturation of 0.2 are shown. This can be
changed with the \code{minsat} argument.

Upon exit, a code fragment containing a vector with the chosen
color names is placed in the clipboard. It can then be pasted into
code.
}
\examples{
\dontrun{
pick_color()
pick_color("olive")
}
}
\seealso{
\code{\link{colors}}
}
