% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2D.R
\docType{methods}
\name{plot2D}
\alias{plot2D}
\alias{lines2D,gfx2D-method}
\alias{lines2D}
\alias{lines2D-methods}
\alias{lines2D,gfx2D,ANY-method}
\alias{arrows2D,gfx2D-method}
\alias{arrows2D}
\alias{arrows2D-methods}
\alias{arrows2D,gfx2D,ANY-method}
\alias{rect2D,gfx2D-method}
\alias{rect2D}
\alias{rect2D,gfx2D,ANY-method}
\alias{polygon2D,gfx2D-method}
\alias{polygon2D}
\alias{points2D,gfx2D-method}
\alias{points2D}
\alias{text2D,gfx2D-method}
\alias{text2D}
\alias{text2D,gfx2D,ANY-method}
\alias{circle2D,gfx2D-method}
\alias{circle2D}
\alias{demo2D}
\title{Plotting methods for gfx2D objects}
\usage{
\S4method{lines2D}{gfx2D}(object, x, y = NULL, type = "l", ...)

\S4method{arrows2D}{gfx2D}(object, x0, y0, x1 = x0, y1 = y0,
  length = 0.25, ...)

\S4method{rect2D}{gfx2D}(object, xleft, ybottom, xright, ytop,
  density = NULL, angle = 45, col = NA, border = NULL,
  lty = par("lty"), lwd = par("lwd"), ...)

\S4method{polygon2D}{gfx2D}(object, x, y = NULL, ...)

\S4method{points2D}{gfx2D}(object, x, y = NULL, ...)

\S4method{text2D}{gfx2D}(object, x, y = NULL, labels = seq_along(x$x),
  adj = NULL, pos = NULL, offset = 0.5, vfont = NULL, cex = 1,
  col = NULL, font = NULL, srt = 0, ...)

\S4method{circle2D}{gfx2D}(object, x, y = NULL, r = NULL, rx = r,
  ry = rx, pos = "center", phi = 0, n = 16, density = NULL,
  angle = 45, col = NA, border = NULL, lty = par("lty"),
  lwd = par("lwd"), ...)

demo2D()
}
\arguments{
\item{object}{object of class gfx2D}

\item{x, y}{coordinates}

\item{type}{lines}

\item{...}{extra arguments}

\item{x0, y0, x1, y1}{coordinates}

\item{length}{length of arrow head}

\item{xleft, ybottom, xright, ytop}{corners of rectangle}

\item{density, angle}{density and angle of shading lines (see \code{rect} for details)}

\item{border, col}{border and fill color of object}

\item{lty, lwd}{line type and width of object outline

code{rect2D} draws one or more rectangles, similar to the regular
\code{\link[graphics]{rect}} function.}

\item{labels}{text to write}

\item{adj, pos, offset}{text positioning options, as explained in
detail \code{text}}

\item{vfont, font}{font write text}

\item{cex}{character enlargement; note tht this is the enlargement
that is applied in addition to any scaling specified in the
transformation matrix}

\item{srt}{text rotation in degrees; note that this is the rotation
that is in addition to any rotation specified in the
transformation matrix}

\item{r, rx, ry}{radii of ellipse. If only \code{r} or \code{rx} are
specified, a circle is drawn. If \code{rx} and \code{ry} are
specified, an ellipse with these major radii is drawn.}

\item{phi}{angle by which the ellipse is rotated, in radians}

\item{n}{number of corners of polygon (chose a large number to approximate a circle)}
}
\description{
these mostly are wrappers around corresponding base graphics
calls, with coordinate transformations and parameter adjustments
applied

shrt of circle2d
}
\details{
long of circle2d
}
