% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{plotellipse}
\alias{plotellipse}
\title{Draw confidence ellipse}
\usage{
plotellipse(
  data,
  columns = c(1, 2),
  subset = NULL,
  level = 0.67,
  which = c(1, 2),
  fill.ellipse = NA,
  col.ellipse = "black",
  col.ma = "black",
  lwd.ellipse = 1,
  lwd.ma = 1,
  lty.ellipse = 1,
  lty.ma = 1,
  caps = 0
)
}
\arguments{
\item{data}{data frame or matrix with x and y values}

\item{columns}{columns that hold x and y data
(default 1 and 2; column names can also be passed)}

\item{subset}{logical indicating a subset of the data}

\item{level}{confidence interval to draw (default = 0.67)}

\item{which}{which elements are to be drawn (1: ellipse, 2: main axes)}

\item{fill.ellipse}{color in which ellipse area is filled, or NA for no fill}

\item{col.ellipse}{color of ellipse contour, or NA for no contour}

\item{col.ma}{color of major axes}

\item{lwd.ellipse}{width for ellipse}

\item{lwd.ma}{line width for major axes}

\item{lty.ellipse}{line type of ellipse}

\item{lty.ma}{line type of major axes}

\item{caps}{bar ends added to major axes (if any are plotted, 0 = no ends)}
}
\description{
Draw a confidence ellipse for bivariate datainto a bivariate plot
}
\details{
This function draws a confidence ellipse and/or its major axes into
a bivariate plot.
}
\note{
Part of the code is taken and adapted from the \code{ellipse} package,
    for my own purpouses.
}
\examples{
p <- rnorm(100,sd=1)
d <- data.frame(x=30+p+rnorm(100,sd=1), y=20-.5*p+rnorm(100,sd=2))
e <- ellipse(d)
plot(d$x,d$y,asp=1)
plotellipse(d,caps=.2,col.ma="red",lwd.ma=2,fill="#ffff0040") 
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
\keyword{misc,}
\keyword{utilities}
