% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_cartesian.R
\name{pol2cart}
\alias{pol2cart}
\alias{cart2pol}
\title{Convert between cartesian and polar coordinates}
\usage{
pol2cart(r, phi = NULL, deg = FALSE, cols = c("r", "phi"))

cart2pol(x, y = NULL, deg = FALSE, cols = c("x", "y"))
}
\arguments{
\item{r, phi}{Polar coordinates. The first argument (\code{r}) may
    also be a \code{data.frame}. If this \code{data.frame} contains
    columns named \code{"r"} and \code{"phi"} (or alternative names
    provided using argument \code{cols=...}, then these are
    used. Otherwise, the first and second column in the
    \code{data.frame} are assumed to contain \code{r} and
    \code{phi}, respectively.
#'}

\item{deg}{logical indicating whether angles are passed as radians
(default) or degrees \code{deg = FALSE}.}

\item{cols}{Names of columns in \code{data.frame} that contains the
coordinates. If these names are not found, then the 1st and 2nd
column of the \code{data.frame} are used.}

\item{x, y}{Cartesian coordinates. The first argument (\code{x}) may
also be a \code{data.frame}. If this \code{data.frame} contains
columns named \code{"x"} and \code{"y"} (or alternative names
provided using argument \code{cols=...}, then these are
used. Otherwise, the first and second column in the
\code{data.frame} are assumed to contain \code{x} and
\code{y}, respectively.}
}
\value{
Data frame with polar or cartesian coordinates.
}
\description{
\code{pol2cart} and \code{cart2pol} interconvert between the two
coordinate formates. Angles can be provided as radians (default) or
in degrees (\code{deg = TRUE}).
}
\examples{
pol2cart(2, 270, deg=TRUE)



}
