% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyarea.R
\name{polyarea}
\alias{polyarea}
\title{Area of polygon}
\usage{
polyarea(x)
}
\arguments{
\item{x}{polygon}
}
\value{
The area of the polygon
}
\description{
Calculate area of polygon.
}
\details{
The area of the polygon is calculated using the shoelace method,
after translation towards the origin to improve numeric accurary.
The polygon passed can be a list of subpolygons.

Simple polygons are lists and must have elements \code{x} and
\code{y}. More complex shapes are lists of polygons. Polygons with
clock-wise orientation have positive area, whereas polygons with
counter-clockwise orientation are holes (negative area). Polygons
cannot be self-intersecting.

These data structures are compatible with the
\code{polyclip} library.
}
\examples{
a <- polyrect(2,3,4,5)
b <- polycircle(2,3,r=1,n=64)
polyarea(a)
polyarea(b)
\dontrun{
library(polyclip)
c <- polyclip(a,b,op="minus")
polyarea(c)
}
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
