% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.R
\name{polycircle}
\alias{polycircle}
\alias{poly2D}
\alias{polystar}
\alias{polyrect}
\title{Generate polygons}
\usage{
polycircle(x, y = NULL, r = NULL, rx = r, ry = rx,
  pos = "center", phi = 0, n = 16)

polystar(x, y = NULL, r = NULL, r1 = r, r2 = r/2, phi = 0,
  n = 5)

polyrect(x, y, w, h, pos = "center", phi = 0)
}
\arguments{
\item{x}{center of polygon, either as vector \code{c(x,y)} or
\code{x} alone if \code{y} is given.}

\item{y}{coordinate
\code{y} of polygon, unless \code{x} is a vector}

\item{r, r1, r2}{radius of circle. For stars, \code{r1} and \code{r2}
indicate the location of the tips and inner nodes of the star.}

\item{rx}{ellipse radius in x direction
(before rotation)}

\item{ry}{ellipse radius in y direction
(before rotation)}

\item{pos}{anchor point of polygon for which the
coordinate (x,y) is given and around which the object is
rotated. Can be any meaningful combination of
"top","bottom","left", and "right". These can also be
abbreviated as "T","B","L","R". Default is "center".}

\item{phi}{angle to rotate in clockwise
direction}

\item{n}{number of points that are used to construct circle or ellipse}

\item{w}{width}

\item{h}{height}
}
\value{
polygon object
}
\description{
Create a polygon with clock-wise orientation
}
\details{
\code{polyrect} creates a rectangle with width \code{w}, height
\code{h}, and origin (\code{x}, \code{y}). The origin can be
anchored to the center of the rectangle (default), or and corner
and side (see argument \code{pos} for details). The rectangle can
be rotated around the origin by angle \code{phi}.

\code{polycircle} creates a regular polygon with \code{n} (default
16) corners. This approximates a circle, but can also be used to
draw regular polygons when a small n is chosen. Chosing radii that
are different in x and y direction results in an ellipse. The
ellipse can be rotated around an anchor. Anchor and origin are
specified as for the \code{polyrect}.  Note that placing an anchor
to a corner refers to the enclosing rectangle, with the anchor
being away from the actual ellipse.

\code{polystar} creates a star with \code{n} tips. Tips and
inter-tip nodes are placed at distance \code{r1} and \code{r2} from
the origin. A rotation by angle \code{phi} can be applied for
proper tip orientation.
 
Polygons are lists with elements \code{x} and \code{y}, and
clock-wise orientation. These data structures are compatible with
the \code{polyclip} library.
}
\examples{
plot(c(),c(),xlim=c(-6,6),ylim=c(-6,6),asp=1,xlab="x",ylab="y")
for(phi in seq(0,2*pi,length=16))
   polygon(polycircle(0,0,rx=3,ry=1,pos="left",phi=phi,n=60),col="#ff000040")
}
\seealso{
affine2D plot2D
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
