% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyarea.R
\name{polycircle}
\alias{polycircle}
\alias{polyrect}
\title{Generate polygon}
\usage{
polycircle(x, y = NULL, r = NULL, rx = r, ry = rx, pos = "center",
  phi = 0, n = 16)

polyrect(x, y, w, h, pos = "center", phi = 0)
}
\arguments{
\item{x}{center of polygon, either as vector \code{c(x,y)} or
\code{x} alone if \code{y} is given.}

\item{y}{coordinate
\code{y} of polygon, unless \code{x} is a vector}

\item{r}{radius of circle}

\item{rx}{ellipse radius in x direction
(before rotation)}

\item{ry}{ellipse radius in y direction
(before rotation)}

\item{pos}{anchor point of polygon for which the
coordinate (x,y) is given and around which the object is
rotated. Can be any meaningful combination of
"top","bottom","left", and "right". These can also be
abbreviated as "T","B","L","R". Default is "center".}

\item{phi}{angle to rotate in clockwise
direction}

\item{n}{number of points that are used to construct circle or ellipse}

\item{w}{width}

\item{h}{height}
}
\value{
polygon object
}
\description{
Create a polygon with clock-wise orientation
}
\details{
The area of the polygon is calculated using the shoelace method,
after translation towards the origin to improve numeric accurary.
The polygon passed can be a list of subpolygons.

Polygons are lists with elements \code{x} and \code{y}, and
clock-wise orientation. These data structures are compatible with
the \code{polyclip} library.
}
\examples{
plot(c(),c(),xlim=c(-6,6),ylim=c(-6,6),asp=1,xlab="x",ylab="y")
for(phi in seq(0,2*pi,length=16))
   polygon(polycircle(0,0,rx=3,ry=1,pos="left",phi=phi,n=60),col="#ff000040")
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
