% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/power_posneg.R
\name{power.posneg}
\alias{power.posneg}
\alias{power.posneg.backtransform}
\title{Power transformation for positive and negative values}
\usage{
power.posneg(x, posexp = 0.5, negexp = 0.5, slope = 1)

power.posneg.backtransform(x, posexp = 0.5, negexp = 0.5, slope = 1)
}
\arguments{
\item{x}{{Data to be transformed}}

\item{posexp}{Exponent for positive x, or \code{"log"} for log transformation.}

\item{negexp}{Exponent for negative x, or \code{"log"} for log transformation.
Note that the function is applied to abs(x), and the sign of
the result inverted.}

\item{slope}{The slope with which the positive and negative transformation
functions join.}
}
\description{
Joins two power or log transformations (one for positive and
one for negative values). The two functions are joined so that they
have a common slope at x=0. This is achieved by applying an x-axis offset.
}
\details{
\code{power.posexp} was written to provide a continuous and differentiable
data transformation function that differs for positive and negative values.
The data transformation applied has the form

\code{y = sign(y) * abs(x-xoffset)^xexp + yoffset},

where \code{xexp} is specified (\code{posexp} or \code{negexp}).
\code{xoffset} is then chosen so that the slope at x=0 equals
the one specified in the function call (\code{slope}).
yoffset if chosen so that y=0 for x=0.
In case of \code{posexp} or \code{negexp = "log"},
the function applied is

\code{y = sign(y) * log(x-xoffset) + yoffset},

with xoffset and yoffset chosen similarly.

\code{power.posexp.backtransform} calculates the inverse of \code{power.posneg}.
}
\note{
This can be a strange transformation. You need to know what you are doing.
}
\examples{
x <- seq(-100,100,length=100)
y<-power.posneg(x,posexp=1/2,negexp=2/3,slope=1)
plot(x,y,ty="l")
x.back<-power.posneg.backtransform(y,posexp=1/2,negexp=2/3,slope=1)
range(x.back-x)
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
\seealso{
\code{\link{mean}} \code{\link{var}} \code{\link{se}} \code{\link{sd}} \code{\link{median}} \code{\link{min}} \code{\link{max}}
}

