% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_csvxz.R
\name{read_partial}
\alias{read_partial}
\alias{read_filtered}
\alias{guess_decompressor}
\title{Read section of a (compressed) csv file}
\usage{
read_partial(file, lines = c(1, NA), decompressor = NULL, debug = FALSE, ...)

read_filtered(
  file,
  condition,
  decompressor = NULL,
  sep = ",",
  debug = FALSE,
  ...
)

guess_decompressor(file)
}
\arguments{
\item{file}{file name}

\item{lines}{line interval to read. The end of the interval may be
\code{NA}, in which case the file is read until the end.}

\item{decompressor}{command to decompress the file. By default, the
decompressor is chosen based on the file ending (.xz, .gz,
.bz2, .zip are currently supported). Parallelized versions
(\code{pigz} and \code{pbzip2}) are chosen if installed.}

\item{debug}{logical (default FALSE). If TRUE, the command executed
is displayed. Useful for debugging}

\item{...}{additional arguments passed to \code{read.csv}.}

\item{condition}{condition specifying which lines are to be
read. Internally, it is passed to \code{awk}. The condition
can reference data columns by their name. Due to the simplified
parsing, fields must be separated by at least one space from
operators. For example, \code{I == 7} will work but not
\code{I==7}. Also, the logical operators must conform to awk's
standards; specifically, one should use the double operators
and not the vectorized counterparts as in \code{R},
e.g. \code{&&} will work but \code{&} will not. See
\url{https://www.gnu.org/software/gawk/manual/html_node/Boolean-Ops.html#Boolean-Ops}
for details}

\item{sep}{character separating fields (defaults to \code{,})}
}
\value{
\code{data.frame} with file contents
}
\description{
The \code{read_partial} function reads part of a (possibly
compressed) csv file. The line range to read is passed with the
\code{lines} argument. \code{read_filtered} is similar but accepts
a \code{condition} that selects the lines to be read based on their
content. \code{guess_decompressor} suggests a command to decompress
the file to be read.
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
