% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reload.pascal.R
\name{reload.pascal}
\alias{reload.pascal}
\title{Reload package}
\usage{
reload.pascal()
}
\description{
\code{reload.pascal} detach package with unloading
}
\details{
Used only by myself
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}

