% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameColumns.R
\name{renameColumns}
\alias{renameColumns}
\title{Rename columns in data.frame}
\usage{
renameColumns(d, cols)
}
\arguments{
\item{d}{\code{data.frame}}

\item{cols}{names of columns, in the form \code{"new=old"}.}
}
\value{
data frame with renamed columns
}
\description{
Rename columns in data frame.
}
\examples{

data(CO2, package="datasets")

d <- renameColumns(CO2,c("type=Type","CO2=conc"))
head(d)
##   Plant   type  Treatment CO2 uptake
## 1   Qn1 Quebec nonchilled  95   16.0
## 2   Qn1 Quebec nonchilled 175   30.4
## 3   Qn1 Quebec nonchilled 250   34.8
## 4   Qn1 Quebec nonchilled 350   37.2
## 5   Qn1 Quebec nonchilled 500   35.3
## 6   Qn1 Quebec nonchilled 675   39.2

}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
