% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resDivider.R
\name{resDivider}
\alias{resDivider}
\title{Find closely matching values for resistor divider}
\usage{
resDivider(ratio, rmin = 1000, rmax = 1e+06, series = pascal::E24,
  exponents = 0:6, n = 2, tol = 0.05, ntop = 10)
}
\arguments{
\item{ratio}{target ratio of lower resistor to total series
resistance.}

\item{rmin}{minimum total series resistance}

\item{rmax}{maximum total series resistance}

\item{series}{resistor series (1 <= values < 10). Defaults to E24
series.  E12 and E48 vectors are available.}

\item{exponents}{defaults to 0:6, specifying values from 1 Ohm
(inclusive) to 10 MOhm (exclusive).}

\item{n}{number of resistors to use. Defaults to 2, but 3 can also
be chosen.}

\item{tol}{acceptable tolerance for ratio. Defaults to 0.1, which
is a maximum of 10\% deviation.}

\item{ntop}{how many solutions (sorted by deviation) are printed}
}
\value{
data table with results
}
\description{
Find values of 2 or 3 resistors that can be used to get as close a
match as possible for a given resister divider ratio. This function
is not much optimized in performance or code-wise so far.  When
three resistors are used, two may be wired in parallel, and the
third in series.
}
\examples{
resDivider(.35,rmin=1000,rmax=3000,n=3,tol=.01)
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
\keyword{misc}
\keyword{utilities}
