% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxmatches.R
\name{rxmatches}
\alias{rxmatches}
\title{Extract Captured Subexpressions from Regular Expressions}
\usage{
rxmatches(x, m)
}
\arguments{
\item{x}{Texts that were matched}

\item{m}{Matches as returned by (g)regexpr}
}
\value{
Data frame containing the matched substrings, or NA if none
    matched. The columns have the corresponding names if named
    subexpressions (\code{(?<name>...)}) were used.
}
\description{
Return captured subexpressions as data frame.
}
\details{
This function is copied with slight modifications from
https://gist.github.com/MrFlick/10413321. It basically returns
subexpressions captured by \code{regexpr} and its relatives.  The
captured subexpressions are returned as a data frame.
}
\examples{
txt <- c("word X123,Y456 xyz","X23 , Y17", "none")
x <- regexpr("^.*X(?<x>\\\\d+) *, *Y(?<y>\\\\d+).*$", txt, perl=TRUE)
rxmatches(txt,x)
##      x     y
## [1,] "123" "456"
## [2,] "23"  "17"
## [3,] NA    NA
}
\seealso{
\code{\link{regexpr}}
}
\author{
MrFlick \url{https://github.com/MrFlick}, copied by Pascal
    Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
