% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/safen.R
\name{safen}
\alias{safen}
\title{Convert factor/character data to numeric values}
\usage{
safen(x)
}
\arguments{
\item{x}{A vector, list or matrix containing character or numeric data, possibly as factor.}
}
\value{
A numeric vector, list or matrix containing the converted data.
}
\description{
\code{safen} will return the numeric representation of data stored in a vector, list, or matrix.
Numeric data are preserved, character data are converted to numeric.
Factors are first converted to character and then to numeric, thus avoiding returning
data that reflects factor levels \code{1,2,3,...} rather than
the numeric value represented by the character string coding for the level.
}
\examples{
library(pascal)
x <- factor(3:5)
x
## [1] 3 4 5
## Levels: 3 4 5
as.numeric(x)
## [1] 1 2 3
safen(x)
## [1] 3 4 5
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}

