% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeSample.R
\name{safeSample}
\alias{safeSample}
\title{safeSample}
\usage{
safeSample(x, size, replace = FALSE, prob = NULL)
}
\arguments{
\item{x}{a vector of one or more elements from which to
choose. Unlike \code{base::sample}, \code{x} is also
interpreted as vector if it is of numeric type.}

\item{size}{a non-negative integer giving the number of items to
choose.}

\item{replace}{should sampling be with replacement?}

\item{prob}{a vector of probability weights for obtaining the
elements of the vector being sampled.}
}
\description{
This function can replace \code{base::sample}, which is ambiguous
for vector lengths of 1 in case the type is numeric.
}
\details{
\code{sample} is IMO unfortunate, because the behaviour of
\code{sample} becomes identical to the one of \code{sample.int}
when the argument \code{x} is of length 1 and of type numeric.
}
\examples{
## these are the same:
x <- 1:10
sample(x, 1)
safeSample(x, 1)

## these are different
x <- c(10)       # a vector of length 1
sample(x, 1)     # will return the same as sample(1:10, 1)
safeSample(x, 1) # will return the value 10
}
\seealso{
\code{\link{sample}}
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
