% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Standard error of a sample}
\usage{
se(x, na.rm = FALSE)
}
\arguments{
\item{x}{vector of data}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
Standard error of sample
}
\description{
This function computes the standard deviation of the mean of the values in \code{x}.
If \code{na.rm} is \code{TRUE}, missing values are removed before computation proceeds.
}
\details{
This convenience function returns \code{sd} divided by the square root
of the number of data points.
}
\examples{
sd(1:5)
## [1] 1.581139
se(1:5)
## [1] 0.7071068
se(c(1:5,NA))
## [1] NA
se(c(1:5,NA),na.rm=TRUE)
## [1] 0.7071068
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
\seealso{
\code{\link{sd}}
}
\keyword{misc}
\keyword{utilities,}

