% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sgnsqrt}
\alias{sgnsqrt}
\title{Square root with sign reconstruction}
\usage{
sgnsqrt(x)
}
\arguments{
\item{x}{values passed to the function}
}
\value{
square root of absolute value of x, with the sign of x reconstructed
}
\description{
Essentially, this function is implemented as \code{sign(x)*sqrt(abs(x))}.
}
\examples{
sgnsqrt(c(16,-16))
## [1]  4 -4
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
\keyword{misc,}
\keyword{utilities}

