% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showSymbols.R
\name{showSymbols}
\alias{showSymbols}
\title{Show base plotting symbols}
\usage{
showSymbols()
}
\description{
This command produces a simple plot showing the symols
corresponding to \code{pch} values of 0 to 25.
}
\examples{
showSymbols()
}
\seealso{
\code{\link{plot}}

\code{\link{points}}
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
