% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorted.code.R
\name{sorted.code}
\alias{sorted.code}
\title{Create unique ID from separate parts that are sorted}
\usage{
sorted.code(..., split = NULL, collapse = NULL)
}
\arguments{
\item{...}{data to be combined, in one of the forms outlined above}

\item{split}{character separating parts, if parts already combined in strings}

\item{collapse}{character used to merge parts after sorting}
}
\value{
vector of strings with merged parts
}
\description{
\code{sorted.code} constructs IDs from individual parts, making
sure that these parts sorted.code(c("a:b:c","z:x:y","b:a"),split=":",collapse="-") # case 3 are sorted in alphabetical order before being
mergheses/Adele Ferrari/ed. This can be used e.g. to create unambiguous label for
community compositions, given sets of species.
}
\details{
The parts to be merged can be passed in four different forms:
\enumerate{
   \item as elements in separate (parallel) vectors.
   \item as a matrix with parts to be combined in rows.
   \item as string, separated by a character specified in \code{split}.
   \item as list of vectors. 
}
The two last forms are useful if variable numbers of parts are to be combined.
}
\examples{
v1 <- sample(LETTERS,10)
v2 <- sample(LETTERS,10)
m <- cbind(v1,v2)
sorted.code(v1,v2,collapse="-")   # case 1
sorted.code(m,collapse="-")       # case 2
sorted.code(c("a:b:c","z:x:y","b:a"),split=":",collapse="-") # case 3 
sorted.code(list(c("a","b","c"),c("z","y")),collapse="-")    # case 4
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}

