% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splittxt.R
\name{splittxt}
\alias{splittxt}
\title{Split (wrap) text strings to given width}
\usage{
splittxt(x, n = 80)
}
\arguments{
\item{x}{Text string to wrap. Can also be a list or vector, in
which case the result is a list of vectors with the wrapped
parts.}

\item{n}{Width of the result in characters (defaults to 80)}
}
\value{
Vector with text split into parts of size n, or a list of
    vectors if several arguments were passed.
}
\description{
\code{splittxt} splits a given text string into parts of a given length,
ignoring word boundaries, whitespace etc. It was originally written to
wrap DNA sequences to a certain line width.
}
\details{
The parameter \code{n} indicated the width of the output. If
\code{x} has length > 1 (e.g. is a list), then a list is returned
with each element of \code{x} split separately.
}
\examples{
splittxt(paste(LETTERS[1:26],collapse=""),10)

}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
