% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suc.R
\name{suc}
\alias{suc}
\title{Unique values, sorted as character}
\usage{
suc(x)
}
\arguments{
\item{x}{Data vector, typically a factor}
}
\description{
This is a convenience function that returns the realised levels of a factor, as character strings, in ascending order.
}
\details{
Basically, I often used loops of the form \code{for(a in sort(unique(as.character(some_factor)))) {...}},
which I found rather painful to write. This can now be abbreviated as \code{for(a in suc(some_factor)) {...}}.
Note that the result is different from \code{levels{some_factor}}, since levels returns all possible levels
a factor can take, even if there are no corresponding values in the vector (e.g. due to subsetting).
}
\examples{
data(CO2)
levels(CO2$Plant)
## [1] "Qn1" "Qn2" "Qn3" "Qc1" "Qc3" "Qc2" "Mn3" "Mn2" "Mn1" "Mc2" "Mc3" "Mc1"
d <- CO2[1:14,]
suc(d$Plant)
## [1] "Qn1" "Qn2"
levels(d$Plant)
## [1] "Qn1" "Qn2" "Qn3" "Qc1" "Qc3" "Qc2" "Mn3" "Mn2" "Mn1" "Mc2" "Mc3" "Mc1"
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
\keyword{misc}
\keyword{utilities,}

