% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test.asreml.R
\name{test.asreml}
\alias{test.asreml}
\title{Displays ANOVA table, stratum variances, and variance components}
\usage{
test.asreml(d.asr, returnWald = FALSE, silent = FALSE)
}
\arguments{
\item{d.asr}{asreml object, returned from calling \code{asreml} or \code{asreml.nvc}}

\item{returnWald}{logical. If TRUE, \code{test.asreml} will return the table with the wald tests.}

\item{silent}{logical. If TRUE, no output will be produced. This is useful when capturing the wald tests for further processing.}
}
\description{
Given an ASReml object, this function displays the corresponding ANOVA table, error strata, and variance components.
}
\details{
Wald tests are reported with the \code{default} denominator degrees of freedom option of \code{wald.asreml}, adjusted according to Kenward \& Roger.
The function tests for convergence of the model. Stratum variances are reported if available in the summary object \code{asreml} provides.
}
\examples{
\dontrun{
library(pascal)
library(asreml)
data(oats, package="asreml")
oats.asr <- asreml(yield~Blocks+Nitrogen*Variety,random=~Blocks:Variety,data=oats)
test.asreml(oats.asr)
}
## ---- Wald tests:
##                  Df denDF F.inc      Pr
## (Intercept)       1    10  1294 6.5e-12 ***
## Blocks            5    10     5   0.012 *
## Nitrogen          3    45    38 2.5e-12 ***
## Variety           2    10     1   0.272
## Nitrogen:Variety  6    45     0   0.932
##
## ---- Stratum variances:
##                df Variance Blocks:Variety R!variance
## Blocks:Variety 10    601.3              4          1
## R!variance     45    177.1              0          1
##
## ---- Variance components:
##                            gamma component std.error z.ratio constraint
## Blocks:Variety!Blocks.var 0.5989     106.1     67.88   1.563   Positive
## R!variance                1.0000     177.1     37.33   4.743   Positive
##
## ---- Dispersion:
## 13.31
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
\seealso{
wald.asreml
}
\keyword{misc}
\keyword{utilities,}

