% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trim.ws}
\alias{trim.ws}
\title{Trims and white space at ends of string}
\usage{
trim.ws(x)
}
\arguments{
\item{x}{character vector with data to trim}
}
\value{
character vector with trimmed data
}
\description{
Removes whitespace and tabulator characters at end of strings
}
\examples{
trim.ws("  hello!  \\t")
## [1] "hello!"
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}

