% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.crossprod.R
\name{vector.crossprod}
\alias{vector.crossprod}
\title{Vector cross product}
\usage{
vector.crossprod(a, b)
}
\arguments{
\item{a}{first 3D-input vector}

\item{b}{second 3D-input vector}
}
\value{
cross-product of \code{a x b}
}
\description{
Calculate the vector cross-product of two 3-D vectors.
}
\details{
This function calculates the cross product of two vectors in
three-dimensional Euclidean space.  The output vector is
perpendicular to the input vectors \code{a} and \code{b} and has a
magnitude equal to the area of the parallelogram spanned by
\code{a} and \code{b}.

Note that \code{crossprod} from package \code{base} calculates the
matrix cross-product, which is different.
}
\examples{
a <- c(1,0,0)
b <- c(0,1,0)
vector.crossprod(a,b)
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
\keyword{misc,}
\keyword{utilities}

