% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.norm.R
\name{vector.norm}
\alias{vector.norm}
\title{p-norm of a vector}
\usage{
vector.norm(x, p = 2)
}
\arguments{
\item{x}{input vector, or list of vectors}

\item{p}{type of norm (default 2)}
}
\value{
norm of vector, or vector with norms of vectors passed in list
}
\description{
Calculate the p-norm of a vector, or of a list of vectors
}
\details{
This function calculates the p-norm of a vector or a list of
vectors, which is \code{(sum_i x_i^p)^(1/p)}. For p=1, this is
equivalent to the Manhattan norm, for p=2 to the Euclidean
norm.

The Euclidean norm of a vector can also be obtained by calling
\code{norm(x,"2")} from package \code{base}, but this is much
slower.
}
\examples{
a <- c(1,2,3)
d <- list(a,c(2,3,4,5,6))
vector.norm(a,1)
vector.norm(a,2)
norm(a,"2")
vector.norm(d)
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
\seealso{
norm
}
\keyword{misc,}
\keyword{utilities}

