% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpand.R
\name{xpand}
\alias{xpand}
\title{Pads a data.frame with additional rows}
\usage{
xpand(
  formula,
  data,
  verbose = FALSE,
  cols = NULL,
  unit = NULL,
  unname = FALSE,
  expand = FALSE
)
}
\arguments{
\item{formula}{Right-sided formula indicating the columns that
specify the groups.}

\item{data}{\code{data.frame} with the data set to expand.}

\item{verbose}{logical indicating whether some indications about
data format and progress should be printed (default = FALSE)}

\item{cols}{string vector with names of columns to include in
result, in addition to the ones specified in the formula
(default = all).}

\item{unit}{string with name of the column that holds the consecutive "units"
in the groups (default = none)}

\item{unname}{logical indicating whether row names referring to the
original \code{data.frame} should be kept (default = TRUE).}

\item{expand}{logical indicating whether all combination of terms
should be included in the result, even if these do not occur in
the original data (default = FALSE).}
}
\value{
Data frame containing the expanded data
}
\description{
\code{xpand} adds additional rows to a \code{data.frame} until the
groups specified have the same size.  This makes using
\code{asreml}'s outer product notation for residual correlations
easier. The added rows all contain \code{NA} values, except for
the key columns defining the groups.
}
\details{
\code{asreml} requires a data frame with complete and equally sized
groups, should a residual covariance matrix be defined with the
outer product notation (options \code{rcov=~...} in ASReml3, and
\code{residuals=~...} in ASReml 4). While this notation is
convenient, padded 'unbalanced' data sets to the required format is
cumbersome.

\code{xpand} does exactly this. \code{formala} is a right-hand side
only formula that contains the group-defining terms. When
\code{expand = TRUE}, all combinations of term values will be
included in the result, even if some groups are missing in the
original data frame.
}
\examples{
data(CO2, package="datasets")
d <- CO2[c(1:5,31:32,48),]
##    Plant        Type  Treatment conc uptake
## 1    Qn1      Quebec nonchilled   95   16.0
## 2    Qn1      Quebec nonchilled  175   30.4
## 3    Qn1      Quebec nonchilled  250   34.8
## 4    Qn1      Quebec nonchilled  350   37.2
## 5    Qn1      Quebec nonchilled  500   35.3
## 31   Qc2      Quebec    chilled  250   35.0
## 32   Qc2      Quebec    chilled  350   38.8
## 48   Mn1 Mississippi nonchilled  675   32.4

xpand(~Type+Treatment, d)
##             Type  Treatment Plant conc uptake
## 1         Quebec nonchilled   Qn1   95   16.0
## 2         Quebec nonchilled   Qn1  175   30.4
## 3         Quebec nonchilled   Qn1  250   34.8
## 4         Quebec nonchilled   Qn1  350   37.2
## 5         Quebec nonchilled   Qn1  500   35.3
## 31        Quebec    chilled   Qc2  250   35.0
## 32        Quebec    chilled   Qc2  350   38.8
## NA        Quebec    chilled  <NA>   NA     NA
## NA.1      Quebec    chilled  <NA>   NA     NA
## NA.2      Quebec    chilled  <NA>   NA     NA
## 48   Mississippi nonchilled   Mn1  675   32.4
## NA.3 Mississippi nonchilled  <NA>   NA     NA
## NA.4 Mississippi nonchilled  <NA>   NA     NA
## NA.5 Mississippi nonchilled  <NA>   NA     NA
## NA.6 Mississippi nonchilled  <NA>   NA     NA

xpand(~Type+Treatment, d, expand = TRUE)
##              Type  Treatment Plant conc uptake
## NA    Mississippi    chilled  <NA>   NA     NA
## NA.1  Mississippi    chilled  <NA>   NA     NA
## NA.2  Mississippi    chilled  <NA>   NA     NA
## NA.3  Mississippi    chilled  <NA>   NA     NA
## NA.4  Mississippi    chilled  <NA>   NA     NA
## 31         Quebec    chilled   Qc2  250   35.0
## 32         Quebec    chilled   Qc2  350   38.8
## NA.5       Quebec    chilled  <NA>   NA     NA
## NA.6       Quebec    chilled  <NA>   NA     NA
## NA.7       Quebec    chilled  <NA>   NA     NA
## 48    Mississippi nonchilled   Mn1  675   32.4
## NA.8  Mississippi nonchilled  <NA>   NA     NA
## NA.9  Mississippi nonchilled  <NA>   NA     NA
## NA.10 Mississippi nonchilled  <NA>   NA     NA
## NA.11 Mississippi nonchilled  <NA>   NA     NA
## 1          Quebec nonchilled   Qn1   95   16.0
## 2          Quebec nonchilled   Qn1  175   30.4
## 3          Quebec nonchilled   Qn1  250   34.8
## 4          Quebec nonchilled   Qn1  350   37.2
## 5          Quebec nonchilled   Qn1  500   35.3

}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
