% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsink.R
\name{Sink}
\alias{Sink}
\alias{enableSink}
\alias{disableSink}
\alias{checkSink}
\title{Sink that can be enabled/disabled}
\usage{
Sink(...)

enableSink(enabled = TRUE)

disableSink()

checkSink()
}
\arguments{
\item{...}{parameters passed to \code{sink}}

\item{enabled}{TRUE to enable and FALSE to disable sinking.}
}
\value{
\code{checkSink()} will return TRUE if sinking is enabled, FALSE otherwise.
}
\description{
Often, one sinks results to a file. During debugging, this may be
disadvantageous because the output and important messages may be invisible on the console.
Combining each sink command with a separate \code{if(sinkEnabled) { ... }} appears cumbersome,
and so does commenting out the respective lines.
\code{Sink}, \code{enableSink}, and \code{enableSink} offer alternatives.
Essentially, the \code{Sink(...)} command is only effective if enabled.
}
\examples{
enableSink()
# replacing the above line with enableSink(FALSE) or disableSink() will
# turn off the effect of the following Sink commands
Sink("file.txt")
# do some stuff here
Sink()
}
\seealso{
\code{\link{sink}}
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
