% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xtreme.R
\name{xtreme}
\alias{xtreme}
\title{Robust detection of extreme values}
\usage{
xtreme(x, f = 3.5)
}
\arguments{
\item{x}{vector of data}

\item{f}{number of sd-equivalents defining extreme values}
}
\description{
Detect extreme values based on interval defined by median absolute deviation
}
\details{
The function \code{xtreme} checks which values are more than \code{f} median absolute deviations
(scaled to standard deviation equivalents) away from the median.
With the default value of f = 3.5 and a perfect normal distribution, only 1 in ca. 2000
values would be classified as extremes. With f = 4, the fraction of extremes reduces to 1 in 16000 values.
}
\examples{
pnorm(-3.5)*2; # fraction of data more than 3.5 sd away from mean
y <- rnorm(1e6)
sum(xtreme(y))/1e6;
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
\seealso{
\code{\link{mad}}
}
\keyword{misc}
\keyword{utilities}

