% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy.errbar.R
\name{xy.errbar}
\alias{xy.errbar}
\title{Plot error bars}
\usage{
xy.errbar(x, y, yplus = NULL, yminus = NULL, xplus = NULL,
  xminus = NULL, yerr = NULL, xerr = NULL, cap = 0.015, lty = 1,
  lwd = 1, add = FALSE, ...)
}
\arguments{
\item{x}{vector of x coordinates of data points to be fitted with error bars}

\item{y}{vector of y coordinates of data points to be fitted with error bars}

\item{yplus}{extention of y error bar in positive (upward) direction}

\item{yminus}{extention of y error bar in negative (downward) direction (this is a positive number)}

\item{xplus}{extention of x error bar in positive (rightward) direction}

\item{xminus}{extention of x error bar in negative (leftward) direction (this is a positive number)}

\item{yerr}{alternative notation for symmetric error bars: if given, then yplus and yminus are set to this value}

\item{xerr}{alternative notation for symmetric error bars: if given, then xplus and xminus are set to this value}

\item{cap}{width of the little lines at the end of the error bars in units of the width of the plot. Default is .015.}

\item{lty}{line type for error bars}

\item{lwd}{line width for error bars}

\item{add}{set to TRUE to add bars to an existing plot}

\item{...}{extra arguments passed to \code{segments} , e.g. to colors}
}
\description{
Given a set of points (x,y), error bars in horizontal and/or vertical direction are plotted.
}
\examples{
## example: barplot
data(CO2, package="datasets")
CO2.aggr<-aggr(CO2,c("Treatment"),c("uptake=mean(uptake)","se.uptake=se(uptake)"))
x<-barplot(CO2.aggr$uptake,names.arg=CO2.aggr$Treatment,ylim=c(0,max(CO2.aggr$uptake+CO2.aggr$se.uptake)))
xy.errbar(x,CO2.aggr$uptake,yerr=CO2.aggr$se.uptake,add=TRUE)

## example: x/y plot
CO2.aggr<-aggr(CO2,c("conc"),c("uptake=mean(uptake)","se.uptake=se(uptake)"))
xy.errbar(safen(CO2.aggr$conc),CO2.aggr$uptake,yerr=CO2.aggr$se.uptake)
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}

