% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Zr.R
\name{Zr}
\alias{Zr}
\title{Calculate normalised effect sizes}
\usage{
Zr(F, dendf, df = 1)
}
\arguments{
\item{F}{F-ratio (e.g. from ANOVA table)}

\item{dendf}{denominator degrees of freedom}

\item{df}{nominator degrees of freedom (default: 1)}
}
\value{
Fisher's Z-transformed effect size
}
\description{
In meta-analyses, one often uses Z-transformed effect sizes.  This
function returns such effect sizes, computed from F values and
associated nominator and denominator degrees of freedom.
}
\examples{
Zr(4.5,10)
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}

