% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addColumns.R
\name{addColumns}
\alias{addColumns}
\title{Add columns to data.frame}
\usage{
addColumns(trg, src, key, cols, after = TRUE)
}
\arguments{
\item{trg, src}{target and source \code{data.frame}s}

\item{key}{key column(s). Either a single name, in which case it
applies to both source and target, or a vector of two, with
first and second applying to target and source, respectively.}

\item{cols}{names of columns to copy to target. The columns can be
renamed by using the form \code{"new=old"}.}

\item{after}{logical. Indicated whether the new columns are placed
right after the key column (default, after=TRUE), or before the
key column (adter=FALSE). The new columns always are in the
order in which they were specified.}
}
\value{
target data frame amended with columns.
}
\description{
Adds columns from another data.frame based on a key.
}
\details{
This is useful when In meta-analyses, one often uses Z-transformed effect sizes.  This
function returns such effect sizes, computed from F values and
associated nominator and denominator degrees of freedom.
}
\examples{

data(CO2, package="datasets")

d <- aggr(CO2,"Plant", "mean.uptake=mean(uptake)")
addColumns(d, CO2, "Plant",c("Type","Treatment"))

## differnet key columns, and column renaming
d <- aggr(CO2,"p=Plant", "mean.uptake=mean(uptake)")
addColumns(d, CO2, c("p","Plant"),c("name=Type","chilled=Treatment"))

}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
