% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine2D.R
\name{translate2D}
\alias{translate2D}
\alias{affine2D}
\alias{scale2D}
\alias{rotate2D}
\alias{shearx2D}
\alias{sheary2D}
\alias{reflectxy2D}
\alias{reflectx2D}
\alias{reflecty2D}
\alias{identity2D}
\title{Matrices for affince 2D transformations}
\usage{
translate2D(x, y)

scale2D(kx, ky = kx)

rotate2D(theta)

shearx2D(phi)

sheary2D(phi)

reflectxy2D()

reflectx2D()

reflecty2D()

identity2D()
}
\arguments{
\item{x}{amount to translate along x axis}

\item{y}{amount to translate along y axis}

\item{kx}{scale factor along x axis}

\item{ky}{scale factor along y axis}

\item{theta}{angle to rotate, clock-wise}

\item{phi}{angle to shear, away from other axis in direction of 1st quadrant}
}
\value{
3 x 3 transformation matrix
}
\description{
Create 3 x 3 matrices for affine transformations
}
\examples{
p1 <- polycircle(0,0,rx=3,ry=1,pos="BL")
plot(NULL,xlim=c(-5,5),ylim=c(-5,5),asp=1,xlab="",ylab="")
polygon(p1,border="red")
m <- scale2D(1.5) \%*\% rotate2D(pi/6) \%*\% translate2D(-3,-1)
p2 <- apply2D(m,p1)
polygon(p2,border="blue")
}
\seealso{
plot2D
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
