% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binvar.R
\name{binvar}
\alias{binvar}
\title{Variance of distribution of binned data}
\usage{
binvar(x, rho, width = 1)
}
\arguments{
\item{x}{vector with the center of the bins}

\item{rho}{relative probability density for each bin. \code{rho} is
    normalized to unity integral, i.e. \code{sum(width*rho_i) = 1}.}

\item{width}{width of the bin}
}
\value{
variance of distribution
}
\description{
Calculates the population variance of a distribution given by bins
of a certain size, assuming an uniform distribution within each bin,
i.e. a ``staircase-shaped'' probability density function.
Note that the estimated variance will be biased in the case of discretized data,
i.e. it will differ from the one of the original population.
}
\examples{
-4
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
\keyword{misc,}
\keyword{utilities}

