% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloglog.R
\name{cloglog}
\alias{cloglog}
\alias{invcloglog}
\title{Complementary log-log link function and its inverse}
\usage{
cloglog(x)

invcloglog(x)
}
\arguments{
\item{x}{Numeric vector or matrix containing data to be transformed}
}
\value{
Transformed data
}
\description{
\code{cloglog} will return the complementary log log function, whereas
\code{invcloglog} will invert it.
\code{cloglog} will yield -Inf and Inf for arguments 0 and 1, or NaN for values
outside this range. Conversely, \code{invcloglog} handles inputs of -Inf and Inf.
}
\examples{
cloglog(.3)
## [1] -1.03093

invcloglog(cloglog(.3))
## [1] 0.3

}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}

