% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplots.R
\name{groupSpace}
\alias{groupSpace}
\title{Add spacing between bars in bar plot}
\usage{
groupSpace(d, cols)
}
\arguments{
\item{d}{data frame with data}

\item{cols}{the names of the columns that define the groups.}
}
\value{
A vector with values of 1 when a transition occurs.
}
\description{
\code{groupSpace} is a helper function to add space between the
bars of a bar plot.
}
\details{
Given a \code{data.frame} and the names of columns in it, it
determines transitions between the groups these columns define and
returns 1 for a transition, or 0 otherwise. These values can be
passed to the \code{space} argument of \code{barplot}.
}
\examples{
\dontrun{



library(grid) # required only because we use grid.cirle in fun
plot(rnorm(10),rnorm(10))
corner.label("a",pos="topleft",dist=2,cex=2)
corner.label(label="b",pos="topright",
             fun=function(x,y,...)
                 grid.circle(x,y,unit(.7,"char"),
                             gp=gpar(fill="black",...)),
             dist=2,units="cm",
             col="white",
             fontface="bold",
             fontsize=30)
corner.label("c",pos="bottomright",cex=2)
corner.label("ddd",pos="bottomleft",dist=c(1.7,1),cex=2)
}
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
