% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/escape_unescape.R
\name{unescape}
\alias{unescape}
\title{Unescape special sequences in strings}
\usage{
unescape(x)
}
\arguments{
\item{x}{Vector, list or matrix of character strings with escaped special characters}
}
\value{
Vector, list or matrix of strings with the corresponding sequences unescaped
}
\description{
This is the inverse function of escape.
}
\examples{
r<-escape(list("A:B","C:D"))
r
## [[1]]
## [1] "A\\\\003aB"
##
## [[2]]
## [1] "C\\\\003aD"
unescape(r)
## [[1]]
## [1] "A:B"
##
## [[2]]
## [1] "C:D"
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
\seealso{
\code{\link{escape}}
}

