% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Zr.R
\name{Zr}
\alias{Zr}
\title{Calculate normalised effect sizes}
\usage{
Zr(F, dendf, df = 1)
}
\arguments{
\item{aovobj}{An object containing the result of \code{aov}}

\item{test.formula}{A formula specifying nominator and denominator of the test.
Several formula can be passed as a list.}

\item{table}{TRUE indicates that the results should be returned as a formatted table,
whereas FALSE will return a vector of P values.}
}
\value{
A vector of P-values if \code{table=FALSE}, or a \code{data.frame} containing
a formatted ANOVA table if \code{table=TRUE}. Note that all P values are rounded
up to the next displayed digit, i.e. all P values < 0.001 will be indicated as 0.001.
}
\description{
Calculates effects sizes from F values and associated denominator
degrees of freedom.
}
\examples{
library(pascal)

Zr(4.5,10)
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
\seealso{
\code{\link{aov}}
}

